test: rm reth-pirmitives block tests
